Shelve everything
